fix(encryption): harden persistence and provider boundaries
- Add `insert_encrypted`, `update_encrypted`, and `save_encrypted` on
  `Encryptable`: run `before_save`, encrypt, write, run `after_save`.
  `encrypt_fields_ctx` stays for custom writes such as upserts.
- `ModelDecryption` is implemented for every `ModelTrait`; `decrypt_fields_ctx`
  and `decrypt_fields` read and write columns through `ModelTrait::get` /
  `try_set`, so the entity turbofish and the serde round-trip are gone.
- `row_scope_from_json` becomes `row_scope_from_model`; scope values come
  from `sea_orm::Value` on both paths. Bytes and array columns are rejected.
- The envelope key id `h.i` is part of the AES-GCM associated data. The
  config provider uses the first 8 hex characters of SHA-256 over the key as
  the id. Decryption tries the key that matches `h.i` first.
- `encrypt_compressed` leaves values above `MAX_INFLATED_BYTES` uncompressed.
- `EncryptionConfig` redacts keys in `Debug` and `Serialize` and zeroizes on
  drop; boot moves the config out of `ctx.config` after registration.
- `Error::Encryption` carries `EncryptionError`; controllers map it to 500.
- `cipher` and `format` modules are private; `COMPRESS_THRESHOLD` is
  re-exported. `EncryptionError` is `#[non_exhaustive]`.
- The macro rejects duplicate fields and encrypted/scope overlap at compile
  time. `looks_like_envelope` parses the JSON before accepting a value.
- Example and testing helpers use `i64` ids.

+++ src/testing/encryption.rs
@@ -72,7 +72,7 @@ pub async fn ctx_with_encryption_config(cfg: &EncryptionConfig) -> AppContext {
 pub async fn raw_string_column(
     db: &DatabaseConnection,
     table: &str,
-    id: i32,
+    id: i64,
     column: &str,
 ) -> String {
     raw_nullable_column(db, table, id, column)
@@ -87,7 +87,7 @@ pub async fn raw_string_column(
 pub async fn raw_nullable_column(
     db: &DatabaseConnection,
     table: &str,
-    id: i32,
+    id: i64,
     column: &str,
 ) -> Option<String> {
     let backend = db.get_database_backend();
